addressed most review comments, added save cancel button on edit, save cohort, cohort name done
