+++ visualization/dashboard/src/MLIDashboard/JointDataset.ts
@@ -176,8 +176,6 @@ export class JointDataset {
                 sortedCategoricalValues: args.metadata.modelType !== ModelTypes.regression ? args.metadata.classNames : undefined,
                 category: ColumnCategories.outcome
             };
-            const predictions = args.predictedY
-            const average = predictions.reduce((a, b) => a + b) / predictions.length;
             if (args.metadata.modelType === ModelTypes.regression) {
                 this.metaDict[JointDataset.PredictedYLabel].featureRange = {
                     min: Math.min(...args.predictedY),
