+++ visualization/mlchartlib/src/INumericRange.ts
@@ -3,7 +3,6 @@ import { RangeTypes } from "./RangeTypes";
 export interface INumericRange {
     // if the feature is numeric
     min: number;
-    avg:number;
     max: number;
     rangeType: RangeTypes.integer | RangeTypes.numeric;
 }
